LayoutTests/fast/forms/input-maxlength-ime-preedit-expected.txt

This test checks users can input pre-edit text longer than maxlength.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.selectionStart is 0
PASS input.selectionEnd is 4
PASS input.value.substring(input.selectionStart, input.selectionEnd) is "abcd"
PASS successfullyParsed is true

TEST COMPLETE
